Extend tests and address pr comments
