+++ mapproxy/test/system/fixture/dimension.yaml
@@ -1,29 +1,52 @@
 services:
   demo:
   wms:
-    srs: ['EPSG:4326', 'EPSG:3857']
-    bbox_srs: ['EPSG:4326', 'EPSG:3857']
-    versions: ['1.0.0','1.1.0','1.1.1', '1.3.0' ]
+    image_formats: ['image/png', 'image/jpeg', 'png8', 'image/tiff']
+    srs: ['EPSG:4326', 'EPSG:900913', 'EPSG:3857']
+    bbox_srs: ['EPSG:4326', 'EPSG:900913', 'EPSG:3857']
+    versions: ['1.0.0','1.1.0','1.1.1', '1.3.0' ]   
     md:
       title: Test Dimension
 caches:
   test_cache:
-    sources: [test]
-    grids: [GLOBAL_GEODETIC]
+    format: image/jpeg
+    sources: [test_cache]
 
 sources:
-  test:
+  test_cache:
     type: wms
+    wms_opts:
+      featureinfo: True
     req:
-      url: https://example.url/geomet/?
-      layers: Layer_Test
+      url: http://localhost:42422/service
+      layers: foo,bar
     forward_req_params: ["time","dim_reference_time"]
 
 globals:
+  http:
+    hide_error_details: false
+  cache:
+    base_dir: cache_data/
+    meta_size: [1, 1]
+    meta_buffer: 0
+    tile_lock_dir: defaulttilelockdir
 
+  srs:
+    preferred_src_proj:
+      'EPSG:25831': ['EPSG:25832', 'EPSG:3857']
+
+  image:
+    # resampling: 'bicubic'
+    paletted: False
+    formats:
+      custom:
+        format: image/jpeg
+      png8:
+        format: 'image/png; mode=8bit'
+        colors: 256
 layers:
-  - name: Layer_Test
-    title: Layer_Test - Dimension
+  - name: test_cache
+    title: Test WMS Cache Layer - Dimension
     sources: [test_cache]
     dimensions:
       time:
@@ -34,4 +57,3 @@ layers:
         values:
           - "2020-09-22T11:20:00Z/2020-09-22T14:20:00Z/PT2H"
         default: "2020-09-22T14:20:00Z"
-
